Add generic error handling
This commit introduces generic error handling in service templates. It
defines an HTTPError object, which can be thrown by clients to signal
errors. Each route handler is modified and wrapped in a promised try()
block which allows catching sync and async errors alike. These are then
handled in a generic error function set with app.use().