Add a docker wrapper script
This utility script builds the image using the provided Dockerfile,
starts the service inside the container and deletes the container after
the end of the execution, i.e. when the user Ctrl-C's it. Additionally,
if the '-t' or the '--test' command line argument is supplied, instead
of running the service, the script runs 'npm test' inside the container.