Attach a request ID to each request
For each request, the middleware now inspects it for the existence of
the 'X-Request-Id' header. If one is not present, a new request ID is
generated and attached to the request.

Additionally, a new child logger object is now created per request,
containing the aforementioned request ID, such that when
`req.logger.log()` is used, the request ID output with the logged
message/object.