WIP: Read zoom and attribution config from /<style>/info.json
This allows different installs to use their own setting while
reducing hard-coding.

This changes the default max zoom (used only if the info.json request
fails) to 18.  This is on the theory that 18 is safe if 20 is
available, but not vice-versa.

This is based off a tag (0.0.17), since that is the only way I can
get it to work with update-snapshot-dep. It needs to be rebased.

Bug: T186780