[WIP] Add Travis-CI support for unit tests
The tests were crashing due to a missing config file for local
testing, and due to not having geoshapes requestHandler.

The consideration for the local requestHandler should be done better
but the tests at least do not explode now. They are "just" all failing
and it's unclear if there's an issue still with the service config
or if the tests are accurately showing a problem with the service.

Continuing to investigate while this is in WIP.