Force linting of es5 file
Since the airbnb-base file is completely configured for es6, even
canceling the environment param 'node' doesn't help, as all the
rules still apply. We have to make sure the browser-facing file is
properly linted to work in supported browsers. For that purpose,
the ruleset for wikimedia is used for the es5 file alone, for the
sake of having a base lint that allows es5 rules.

Fixes #7