Add copy info to initial setup
When loading style from yaml file, it queries the source info
to reconcile the list of layers with those present in the
source. When the keyspace is new and empty, any source that
depends on cassandra (vXview, osm-pbf, etc) fails because
the list of layers from getInfo() is empty.

This change makes copying info from the source (generally 'gen')
into a new cassandra keyspace part of the initial setup activated
by 'createIfMissing: true'. When this options is used, it needs
a new parameter: 'copyInfoFrom: {ref: gen}'

Bug: https://phabricator.wikimedia.org/T192058