Add an event-emitting facility
Adds an EventService class with the initial purpose of emitting resource
change events to prompt tile invalidation in Varnish.  The resource change
events are formatted as specified in the relevant schema in
https://github.com/wikimedia/mediawiki-event-schemas, for example:

  {
    "meta": {
      "domain": "maps.localhost",
      "dt": "2018-08-29T21:17:34.361Z",
      "id": "f26d80a0-abd0-11e8-a2ca-c789637275fa",
      "schema_uri": "resource_change/1",
      "topic": "resource_change",
      "uri": "//maps.localhost/osm-intl/12/2074/1405.png"
    },
    "tags": [
      "tilerator"
    ]
  }

The event processing URI, tile server domain, and sources to invalidate
are to be specified (optionally) in the service configuration file.

Bug: https://phabricator.wikimedia.org/T109776