Fix var initialization in lib/autoPosition.js
A recent change removed the semicolon from the end of the list of
variables initialized at the beginning of the file, introducing a runtime
error.

This change fixes the issue by initializing one variable at a time (in
lines ending in semicolons).  const is used instead of let as it's more
appropriate (these values don't change).